requirements.txt
functions-framework==3.*
pandas==1.5.3
psycopg2-binary==2.9.1
google-cloud-secret-manager==2.4.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
cloud-sql-python-connector==1.2.4
pg8000==1.29.8
sqlalchemy==2.0.0